Get output before setting voltage
Setting voltage implements a protection against erroneous use - when pin
is configured as input it is not allowed to set voltage. The Tango device
uses the last read output to validate this condition. This has two problems.
First, when output was never read it is not known and an unhandled exception
occurs when setting voltage. Second, when the output was changed externally
in between the last output read and the voltage set this protection won't work.
Read output at every voltage set to overcome the first problem and mitigate
the second one.